refactored formValidation.php so equivalent test file not needed
